misc case detail redesign updates
