+++ web-client/src/presenter/computeds/docketRecordHelper.js
@@ -2,14 +2,8 @@ import { state } from 'cerebral';
 
 export const docketRecordHelper = get => {
   const permissions = get(state.permissions);
-  const currentPage = get(state.currentPage);
-  const userAssociatedWithCase = get(state.screenMetadata.isAssociated);
 
   const showDirectDownloadLink = !permissions.UPDATE_CASE;
-  const showFileDocumentButton =
-    permissions.FILE_EXTERNAL_DOCUMENT &&
-    ['CaseDetail'].includes(currentPage) &&
-    userAssociatedWithCase;
 
   const canShowEditDocketEntryLink = permissions.DOCKET_ENTRY;
 
@@ -18,6 +12,5 @@ export const docketRecordHelper = get => {
     showDirectDownloadLink,
     showDocumentDetailLink: !showDirectDownloadLink,
     showEditDocketEntry: permissions.DOCKET_ENTRY,
-    showFileDocumentButton,
   };
 };
