+++ web-client/src/presenter/computeds/docketRecordHelper.test.js
@@ -98,51 +98,6 @@ describe('docket record helper', () => {
     expect(result.showEditDocketEntry).toEqual(false);
   });
 
-  it('should show file document button if user has FILE_EXTERNAL_DOCUMENT permission and the user is associated with the case', () => {
-    const result = runCompute(docketRecordHelper, {
-      state: {
-        caseDetail: {},
-        currentPage: 'CaseDetail',
-        form: {},
-        permissions: {
-          FILE_EXTERNAL_DOCUMENT: true,
-        },
-        screenMetadata: { isAssociated: true },
-      },
-    });
-    expect(result.showFileDocumentButton).toEqual(true);
-  });
-
-  it('should not show file document button if user does not have FILE_EXTERNAL_DOCUMENT permission', () => {
-    const result = runCompute(docketRecordHelper, {
-      state: {
-        caseDetail: {},
-        currentPage: 'CaseDetail',
-        form: {},
-        permissions: {
-          FILE_EXTERNAL_DOCUMENT: false,
-        },
-        screenMetadata: { isAssociated: true },
-      },
-    });
-    expect(result.showFileDocumentButton).toEqual(false);
-  });
-
-  it('should not show file document button if user has FILE_EXTERNAL_DOCUMENT permission but the user is not associated with the case', () => {
-    const result = runCompute(docketRecordHelper, {
-      state: {
-        caseDetail: {},
-        currentPage: 'CaseDetail',
-        form: {},
-        permissions: {
-          FILE_EXTERNAL_DOCUMENT: true,
-        },
-        screenMetadata: { isAssociated: false },
-      },
-    });
-    expect(result.showFileDocumentButton).toEqual(false);
-  });
-
   it('should allow the actionable docket entry link to be displayed for user roles that have the DOCKET_ENTRY permissions', async () => {
     let result = runCompute(docketRecordHelper, {
       state: {
