+++ web-client/src/views/CaseDetail/CaseDetailHeader.jsx
@@ -80,8 +80,21 @@ export const CaseDetailHeader = connect(
 
             {!hideActionButtons && caseDetailHeaderHelper.showExternalButtons && (
               <div className="tablet:grid-col-4">
+                {caseDetailHeaderHelper.showFileDocumentButton && (
+                  <Button
+                    secondary
+                    className="tablet-full-width push-right margin-right-0"
+                    href={`/case-detail/${formattedCaseDetail.docketNumber}/before-you-file-a-document`}
+                    icon="file"
+                    id="button-file-document"
+                  >
+                    File a Document
+                  </Button>
+                )}
+
                 {caseDetailHeaderHelper.showRequestAccessToCaseButton && (
                   <Button
+                    secondary
                     className="tablet-full-width push-right margin-right-0"
                     href={`/case-detail/${formattedCaseDetail.docketNumber}/request-access`}
                     id="button-request-access"
@@ -101,6 +114,7 @@ export const CaseDetailHeader = connect(
 
                 {caseDetailHeaderHelper.showFileFirstDocumentButton && (
                   <Button
+                    secondary
                     className="tablet-full-width push-right margin-right-0"
                     href={`/case-detail/${formattedCaseDetail.docketNumber}/file-a-document`}
                     icon="file"
