+++ web-client/src/views/CaseDetail/PetitionerInformation.jsx
@@ -1,13 +1,10 @@
 import { AddPractitionerModal } from './AddPractitionerModal';
-import { AddRespondentModal } from './AddRespondentModal';
 import { Button } from '../../ustc-ui/Button/Button';
 import { EditPractitionersModal } from './EditPractitionersModal';
-import { EditRespondentsModal } from './EditRespondentsModal';
 import { EditSecondaryContactModal } from '../EditSecondaryContactModal';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { FormGroup } from '../../ustc-ui/FormGroup/FormGroup';
 import { PractitionerExistsModal } from './PractitionerExistsModal';
-import { RespondentExistsModal } from './RespondentExistsModal';
 import { connect } from '@cerebral/react';
 import { sequences, state } from 'cerebral';
 import React from 'react';
@@ -49,7 +46,7 @@ const AddressDisplay = (contact, constants, { nameOverride } = {}) => {
   );
 };
 
-const PartyInformation = connect(
+const PetitionerInformation = connect(
   {
     caseDetailHelper: state.caseDetailHelper,
     caseInformationHelper: state.caseInformationHelper,
@@ -58,11 +55,8 @@ const PartyInformation = connect(
     formattedCaseDetail: state.formattedCaseDetail,
     openAddPractitionerModalSequence:
       sequences.openAddPractitionerModalSequence,
-    openAddRespondentModalSequence: sequences.openAddRespondentModalSequence,
     openEditPractitionersModalSequence:
       sequences.openEditPractitionersModalSequence,
-    openEditRespondentsModalSequence:
-      sequences.openEditRespondentsModalSequence,
     openEditSecondaryContactModalSequence:
       sequences.openEditSecondaryContactModalSequence,
     showModal: state.showModal,
@@ -76,9 +70,7 @@ const PartyInformation = connect(
     form,
     formattedCaseDetail,
     openAddPractitionerModalSequence,
-    openAddRespondentModalSequence,
     openEditPractitionersModalSequence,
-    openEditRespondentsModalSequence,
     openEditSecondaryContactModalSequence,
     showModal,
     updateFormValueSequence,
@@ -203,43 +195,6 @@ const PartyInformation = connect(
       </div>
     );
 
-    const respondentPartyInformation = () => (
-      <div className="grid-container padding-x-0">
-        <div className="grid-row">
-          {formattedCaseDetail.respondents &&
-            formattedCaseDetail.respondents.map((respondent, index) => (
-              <div
-                className={classNames(
-                  'tablet:grid-col-3',
-                  index > 3 && 'margin-top-3',
-                )}
-                key={index}
-              >
-                <address aria-labelledby="respondent-label">
-                  {respondent.name &&
-                    AddressDisplay(
-                      {
-                        ...respondent,
-                        ...respondent.contact,
-                      },
-                      constants,
-                      {
-                        nameOverride: respondent.name,
-                      },
-                    )}
-                </address>
-                {respondent.serviceIndicator && (
-                  <div className="margin-top-4">
-                    <span className="semi-bold">Service: </span>
-                    {respondent.serviceIndicator}
-                  </div>
-                )}
-              </div>
-            ))}
-        </div>
-      </div>
-    );
-
     const practitionerSearch = () => (
       <>
         <div className="grid-col-3 text-right">
@@ -251,7 +206,10 @@ const PartyInformation = connect(
           </span>
         </div>
         <div className="grid-col-3 margin-top-neg-05">
-          <FormGroup errorText={validationErrors.practitionerSearchError}>
+          <FormGroup
+            className="margin-bottom-0"
+            errorText={validationErrors.practitionerSearchError}
+          >
             <form
               className="usa-search"
               onSubmit={e => {
@@ -288,65 +246,6 @@ const PartyInformation = connect(
                 <button
                   className="usa-button"
                   id="search-for-practitioner"
-                  type="submit"
-                >
-                  <span className="usa-search__submit-text">Search</span>
-                </button>
-              </div>
-            </form>
-          </FormGroup>
-        </div>
-      </>
-    );
-
-    const respondentSearch = () => (
-      <>
-        <div className="grid-col-3 text-right">
-          <span
-            className="label margin-right-4 margin-top-05"
-            id="respondent-counsel-search-description"
-          >
-            Add Counsel
-          </span>
-        </div>
-        <div className="grid-col-3 margin-top-neg-05">
-          <FormGroup errorText={validationErrors.respondentSearchError}>
-            <form
-              className="usa-search"
-              onSubmit={e => {
-                e.preventDefault();
-                openAddRespondentModalSequence();
-              }}
-            >
-              <div role="search">
-                <label
-                  className="usa-sr-only"
-                  htmlFor="respondent-search-field"
-                >
-                  Search
-                </label>
-                <input
-                  aria-describedby="respondent-counsel-search-description"
-                  className={classNames(
-                    'usa-input margin-bottom-0',
-                    validationErrors.respondentSearchError &&
-                      'usa-input--error',
-                  )}
-                  id="respondent-search-field"
-                  name="respondentSearch"
-                  placeholder="Enter Bar No. or Name"
-                  type="search"
-                  value={form.respondentSearch || ''}
-                  onChange={e => {
-                    updateFormValueSequence({
-                      key: e.target.name,
-                      value: e.target.value,
-                    });
-                  }}
-                />
-                <button
-                  className="usa-button"
-                  id="search-for-respondent"
                   type="submit"
                 >
                   <span className="usa-search__submit-text">Search</span>
@@ -399,44 +298,15 @@ const PartyInformation = connect(
             </div>
           </div>
         )}
-        {caseDetailHelper.showRespondentSection && (
-          <div className="subsection party-information">
-            <div className="card">
-              <div className="content-wrapper">
-                <div className="grid-row header-row">
-                  <div className="grid-col-6 display-flex" id="secondary-label">
-                    <h3>Respondent Counsel</h3>
-                    {caseInformationHelper.showEditRespondents && (
-                      <Button
-                        link
-                        className="margin-left-205 padding-0 height-3"
-                        id="edit-respondents-button"
-                        onClick={() => openEditRespondentsModalSequence()}
-                      >
-                        <FontAwesomeIcon icon="edit" size="sm" />
-                        Edit
-                      </Button>
-                    )}
-                  </div>
-                  {caseInformationHelper.showAddCounsel && respondentSearch()}
-                </div>
-                {respondentPartyInformation()}
-              </div>
-            </div>
-          </div>
-        )}
         {caseDetailHelper.showEditSecondaryContactModal && (
           <EditSecondaryContactModal />
         )}
         {showModal === 'AddPractitionerModal' && <AddPractitionerModal />}
-        {showModal === 'AddRespondentModal' && <AddRespondentModal />}
         {showModal === 'EditPractitionersModal' && <EditPractitionersModal />}
-        {showModal === 'EditRespondentsModal' && <EditRespondentsModal />}
         {showModal === 'PractitionerExistsModal' && <PractitionerExistsModal />}
-        {showModal === 'RespondentExistsModal' && <RespondentExistsModal />}
       </>
     );
   },
 );
 
-export { AddressDisplay, PartyInformation };
+export { AddressDisplay, PetitionerInformation };
